+++ packages/core/src/components/autosuggest/autosuggest.tsx
@@ -192,7 +192,7 @@ export class Autosuggest {
   showLoading = false;
 
   @State()
-  listItemBlockSize = 0;
+  blockSizeFirst10ListItems = 0;
 
   @State()
   listboxContainerMaxBlockSize = 0;
@@ -214,6 +214,7 @@ export class Autosuggest {
 
   private listbox: HTMLDivElement | undefined;
 
+  // ListboxItems are used for the calculation of the listboxContainerMaxBlockSize
   private listboxItems: HTMLDivElement[] = [];
 
   private listboxId: string = v4();
@@ -270,7 +271,10 @@ export class Autosuggest {
   private text = i18n(() => this.host, translations);
 
   componentDidRender() {
-    this.setListboxContainerMaxBlockSize();
+    // This needs to be debounced to avoid a race condition between render() and componentDidRender() whereby
+    // the height of listitems might alternate between two values due to the (dis)appearing vertical scrollbar of the
+    // scrollable's scroll-container
+    debounce(() => this.setListboxContainerMaxBlockSize(), 150);
   }
 
   connectedCallback() {
@@ -341,20 +345,22 @@ export class Autosuggest {
       return;
     }
 
-    if (this.listboxItems[0] && this.showSuggestions) {
-      this.listItemBlockSize = this.listboxItems[0].getBoundingClientRect().height;
+    if (this.showSuggestions) {
+      this.blockSizeFirst10ListItems = 0;
+      const total = this.listboxItems.length ? Math.min(this.listboxItems.length, maxSuggestionsViewable) : 0;
+      for (let i = 0; i < total; i++) {
+        this.blockSizeFirst10ListItems =
+          this.blockSizeFirst10ListItems + (this.listboxItems[i]?.getBoundingClientRect().height || 0);
+      }
     }
 
     const availableBlockSize = window.innerHeight - this.host.getBoundingClientRect().bottom;
-    const listboxMaxBlockSize =
-      this.listItemBlockSize * maxSuggestionsViewable + 2 * listboxPaddingBlock + 2 * listboxBorderWidth;
+    const listboxMaxBlockSize = this.blockSizeFirst10ListItems + 2 * listboxPaddingBlock + 2 * listboxBorderWidth;
 
-    if (availableBlockSize > this.listItemBlockSize) {
-      if (availableBlockSize < listboxMaxBlockSize) {
-        this.listboxContainerMaxBlockSize = availableBlockSize - 2 * listboxPaddingBlock;
-      } else {
-        this.listboxContainerMaxBlockSize = listboxMaxBlockSize;
-      }
+    if (availableBlockSize < listboxMaxBlockSize || availableBlockSize <= this.blockSizeFirst10ListItems) {
+      this.listboxContainerMaxBlockSize = availableBlockSize - 2 * listboxPaddingBlock;
+    } else {
+      this.listboxContainerMaxBlockSize = listboxMaxBlockSize;
     }
   }
 
@@ -728,7 +734,12 @@ export class Autosuggest {
                       const groupLabelId = v4();
                       return (
                         <div role="group" class="group" aria-labelledby={groupLabelId}>
-                          <div class="group-label" role="presentation" id={groupLabelId}>
+                          <div
+                            class="group-label"
+                            role="presentation"
+                            id={groupLabelId}
+                            ref={(element) => element && this.listboxItems.push(element)}
+                          >
                             {suggestionGroup.groupLabel}
                           </div>
                           {suggestionGroup.suggestions.map((suggestion) => (