+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Fixed
+* Autosuggest: memory leak ([#3085](https://github.com/dso-toolkit/dso-toolkit/issues/3085))
+
 ## 📜 Release 70.0.0 - 2025-04-01
 
 ### Docs